npm: adding eslint-import-resolver-typescript
I don't take random dependencies lightly, but this seems to be a better solution for our linter than adding a rule. This allows for imports on TSX/TS files without using their extension, without necessarily ignoring all files or losing out on intellisense.

What this module exactly does is still fuzzy to me, but according to its documentation, and this tutorial (https://andrebnassis.medium.com/setting-eslint-on-a-react-typescript-project-2021-1190a43ffba), it seems to be the way to go.