Refactor & introduce multiple track file uploads
This commit is a lot of fun! To cleanse some of the code, I've seperated the React.useState() for Uploaded Files into its own file, . This allowed me to reuse helpers and export it alongside files, and setFiles. addFiles and clearFiles are the two helpers I made. Furthermore, I have changed the types to expect arrays of files rather than a single file.

Finally, you will notice that the storyboard works, but doesnt parse the text yet. Thats my next step!