Isolated & fixed problem by watching object-properties
ESLint warns to include the object in dependencies, when the object itself should never change, or always change (based on whether we are updating the object in the useEffect or copying it). That doesn't make sense and causes errors.

For example, watching files where `files = [File1, File2, File3]` won't update when a new File is added to the array, because the array is still the same array as it was before. Its the same entry in memory and same pointer. Alternatively, if we copy the array, its always different in memory.

On the other hand, `files.length` WILL change, especially since the files array only gets reset, or added upon. Its entries never change.